+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adding `Norb/MisplacedBranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
 * Adding `Norb/ActiveRecordNamespaced` cop to ensure ActiveRecord models are namespaced. ([@tleish][])
 * Adding `Norb/ActiveRecordThroughBusiness` cop to ensure direct ActiveRecord calls come from business objects. ([@tleish][])
+* Adding `Norb/MisplacedLogic` cop to ensure code logic is not in the wrong place. ([@tleish][])
 
 
 [@tleish]: https://github.com/tleish